Add place list component and display on home page
